Avoid PHP error when things are weird
Fixes https://github.com/JeroenDeDauw/ParamProcessor/issues/16

Might break something else though... I'm not sure why a non-array
is ending up here, and what the value actually is. Is it null?
It it a single value? Is it a string with multiple values that
should have been split into an array?