Use defined params as filter instead of default params; Fix tests
Using the _default_ parameter names to filter parameters does not make sense.
It would force all parameters to be available as positional parameters. Instead
the _defined_ parameter names have to be used.

Remove some test cases that do not make sense in the context of raw parameters.
Raw parameters cannot distinguish between string and other parameter types.
Everything is considered a string. It does not make sense therefore to test
against the type of a parameter.